feat: standardize tabbed headers and nested headers
